% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrintoutsFromGmail.R
\name{getPrintoutFromGmail}
\alias{getPrintoutFromGmail}
\title{Download CSOB printouts from Gmail}
\usage{
getPrintoutFromGmail(year = 2017)
}
\arguments{
\item{year}{for what year printouts should be downloaded}
}
\value{
A tibble with all transactions
}
\description{
Download CSOB printouts from Gmail
}
\examples{
\dontrun{
getPrintoutFromGmail(year = 2018)
}

}
