% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_CSOBprintout.R
\name{parse_CSOBprintout}
\alias{parse_CSOBprintout}
\title{Read/parse CSOB(bank) printout}
\usage{
parse_CSOBprintout(file)
}
\arguments{
\item{file}{CSOB printout, txt, needs to be UTF-8 with Linux line endings}
}
\value{
A tibble with all transactions
}
\description{
Read/parse CSOB(bank) printout
}
\examples{
\dontrun{
parse_CSOBprintout("printout2017_08.txt")
}

}
